requirements.txt
imageio==2.31.2
lifelines==0.27.8
matplotlib==3.5.3
nibabel==4.0.2
numpy==1.21.6
opencv_python==4.10.0.84
pandas==1.3.5
scikit_learn==1.0.2
scipy==1.7.3
shap==0.42.1
tensorflow_intel==2.10.0
torch==2.7.0
